move test folder inside src
