+++ src/test/unit/api/common/helpers/proxy/setup-proxy.test.js
@@ -1,6 +1,6 @@
-import { config } from '../../../../../../../src/config/index.js'
+import { config } from '../../../../../../config/index.js'
 import { getGlobalDispatcher, ProxyAgent } from 'undici'
-import { setupProxy } from '../../../../../../../src/api/common/helpers/proxy/setup-proxy.js'
+import { setupProxy } from '../../../../../../api/common/helpers/proxy/setup-proxy.js'
 
 describe('setupProxy', () => {
   afterEach(() => {
