+++ src/test/unit/api/common/helpers/secure-context/secure-context.test.js
@@ -1,8 +1,8 @@
 import hapi from '@hapi/hapi'
 
-import { secureContext } from '../../../../../../../src/api/common/helpers/secure-context/index.js'
-import { requestLogger } from '../../../../../../../src/api/common/helpers/logging/request-logger.js'
-import { config } from '../../../../../../../src/config/index.js'
+import { secureContext } from '../../../../../../api/common/helpers/secure-context/index.js'
+import { requestLogger } from '../../../../../../api/common/helpers/logging/request-logger.js'
+import { config } from '../../../../../../config/index.js'
 
 const mockAddCACert = jest.fn()
 const mockTlsCreateSecureContext = jest
